# Speed up DB tests