--- a/internal/database/dbstore_db_test.go
+++ b/internal/database/dbstore_db_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/sourcegraph/sourcegraph/internal/database/dbconn"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
 )
 
@@ -14,7 +15,7 @@ func TestMigrations(t *testing.T) {
 	}
 
 	// Setup a global test database
-	db := dbtesting.GetDB(t)
+	db := dbtest.NewDB(t, "")
 
 	migrate := func() {
 		if err := dbconn.MigrateDB(db, dbconn.Frontend); err != nil {