--- a/internal/database/dbtest/dbtest.go
+++ b/internal/database/dbtest/dbtest.go
@@ -71,6 +71,7 @@ func NewDB(t testing.TB, dsn string) *sql.DB {
 
 	config.Path = "/" + dbname
 	testDB := dbConn(t, config)
+	testDB.SetMaxOpenConns(3)
 
 	t.Cleanup(func() {
 		defer db.Close()