With the perf improvements from #21300, `dbtest.NewDB()` is faster than `dbtesting.GetDB()`, and allows tests to be run in parallel. 

This change converts (almost) all instances of `dbtesting.GetDB()` in `internal/database` to `dbtest.NewDB()`, and sets `t.Parallel()` for tests that don't rely on any global state. 

Additionally, since tests are being run in a more parallel fashion, we were hitting the "too many connections" error more often. I added a max open connections limit of 3 to all database connections created with `dbtest.NewDB()` which seems to have helped. 

This change reduced runtime for `internal/database` to ~50s (from ~80 seconds in #21300, which was already reduced from ~120 seconds). 

```
❯ for i in {1..20} ; do go clean -testcache && go test ./internal/database ; done
ok      github.com/sourcegraph/sourcegraph/internal/database    33.562s
ok      github.com/sourcegraph/sourcegraph/internal/database    51.293s
ok      github.com/sourcegraph/sourcegraph/internal/database    51.237s
ok      github.com/sourcegraph/sourcegraph/internal/database    50.243s
ok      github.com/sourcegraph/sourcegraph/internal/database    48.176s
ok      github.com/sourcegraph/sourcegraph/internal/database    49.913s
ok      github.com/sourcegraph/sourcegraph/internal/database    50.085s
ok      github.com/sourcegraph/sourcegraph/internal/database    50.945s
ok      github.com/sourcegraph/sourcegraph/internal/database    48.659s
ok      github.com/sourcegraph/sourcegraph/internal/database    51.025s
```