@camdencheek 2021-05-25
@mrnugget To answer your question about whether `dbtest.NewDB()` is now faster than `dbtesting.GetDB()`, I think this PR is evidence that it is