@camdencheek 2021-05-25
A little more information on where I'm leaving performance:

On average, a call to `NewDB()` takes 1.25 seconds to return, and adds an additional 0.6 seconds to the test cleanup phase. 

When the tests are run with `-parallel 1`, this changes significantly to 0.31 seconds to create the DB and 0.14 seconds to clean it up once the test completes. (note that it still takes ~30 seconds longer overall to run the tests with no parallel)

These two things together indicate to me that much of the difference is due to contention on the Postgres system catalog from all the create/drop operations. If that's the case, there is still likely some improvement to performance on the table if we maintain a pool of migrated databases and just truncate all the tables before returning them. I don't plan to pursue this further at the moment, but figured it would be interesting context to add to the PR.